requirements.txt
fastapi
uvicorn
numpy
scikit-learn
faiss-cpu
sentence-transformers
torch
